Dedupe `lib` `thornweed` entities for PSX
