+++ config/splat.us.stlib.yaml
@@ -51,6 +51,9 @@ segments:
       - [0x26B0, .data, e_misc_2]
       - [0x2714, .data, e_bloody_zombie]
       - [0x280C, data]
+      - [0x2B80, .data, en_thornweed_corpseweed]
+      - [0x2C38, data] # unused data?
+      - [0x2C48, data]
       - [0x2B938, .rodata, first_c_file]
       - [0x2CE78, .rodata, e_shop]
       - [0x2CF14, .rodata, unk_33EC8]
@@ -64,6 +67,7 @@ segments:
       - [0x2DAD8, .rodata, unk_49870]
       - [0x2DB54, .rodata, e_stage_name]
       - [0x2DB88, .rodata, unk_4C054]
+      - [0x2DC48, .rodata, en_thornweed_corpseweed]
       - [0x2DC84, .rodata, e_skeleton]
       - [0x2DD08, .rodata, unk_4F918]
       - [0x2DDA4, .rodata, prim_helpers]
@@ -91,6 +95,7 @@ segments:
       - [0x4B4E0, c, e_particles]
       - [0x4BD24, c, popup]
       - [0x4C054, c, unk_4C054]
+      - [0x4DB6C, c, en_thornweed_corpseweed]
       - [0x4F094, c, e_skeleton]
       - [0x4F918, c, unk_4F918]
       - [0x5139C, c, prim_helpers]
