+++ config/symbols.us.stlib.txt
@@ -210,3 +210,9 @@ D_801CB6C6 = 0x801D5216;
 D_801CB6C8 = 0x801D5218;
 g_CutsceneFlags = 0x801D5284;
 g_ItemIconSlots = 0x801D528C;
+g_EInitThornweed = 0x801808CC;
+g_EInitCorpseweed = 0x801808E4;
+g_EInitCorpseweedProjectile = 0x801808F0;
+EntityThornweed = 0x801CDB6C;
+EntityCorpseweed = 0x801CDEDC;
+EntityCorpseweedProjectile = 0x801CEC0C;
