+++ src/st/lib/en_thornweed_corpseweed.c
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "lib.h"
+
+#define CORPSEWEED_TPAGE 0x14
+#define CORPSEWEED_PAL 0x252
+#define CORPSEWEED_PROJ_PAL 0x255
+#include "../en_thornweed_corpseweed.h"
